Created ViewService

+++ GroupService/src/main/java/org/example/groupservice/repository/GroupRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 
 @Repository
-public interface GroupRepository extends JpaRepository<Group, Integer>
+public interface GroupRepository extends JpaRepository<Group, Long>
 {
     Optional<Group> findByName(String name);
 }
