+++ GroupService/src/main/java/org/example/groupservice/service/StudentService.java
@@ -1,7 +1,7 @@
 package org.example.groupservice.service;
 
 import lombok.AllArgsConstructor;
-import org.example.groupservice.clients.StudentClient;
+import org.example.groupservice.client.StudentClient;
 import org.example.groupservice.dto.StudentResponse;
 import org.springframework.stereotype.Service;
 
@@ -12,12 +12,12 @@ import java.util.List;
 public class StudentService
 {
     private StudentClient studentClient;
-    public void deleteAllStudents(Integer groupId)
+    public void deleteAllStudents(Long groupId)
     {
         studentClient.deleteAllStudents(groupId);
     }
 
-    public List<StudentResponse> getAllStudents(Integer groupId)
+    public List<StudentResponse> getAllStudents(Long groupId)
     {
         return studentClient.allStudents(groupId);
     }
